拆分类
