+++ BDLauncherCSharp/Data/User32.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+
+using BDLauncherCSharp.Extensions;
+
+namespace BDLauncherCSharp.Data
+{
+    public class User32
+    {
+
+        [DllImport("user32.dll")]
+        public static extern bool EnumDisplaySettings(string deviceName, int modeNum, out DEVMODE devMode);
+
+        [DllImport("user32.dll")]
+        public static extern int GetSystemMetrics(int index);
+    }
+}
